requirements.txt
Django==1.6.5
Pillow==2.7.0
PyJWT==0.4.1
South==1.0
argparse==1.2.1
cssmin==0.2.0
cssselect==0.9.1
-e git://github.com/yceruto/django-ajax@23fb813c1216b195716f2ab2c6e8e23c983ec796#egg=django_ajax-dev
django-assets==0.10
django-debug-toolbar==1.2.1
django-debug-toolbar-mongo==0.1.10
django-extensions==1.4.0
django-live-profiler==0.0.9
django-mongonaut==0.2.21
django-simple-captcha==0.4.2
ipdb==0.8
ipython==2.2.0
jsmin==2.0.11
lxml==3.4.0
mongoadmin==0.2
mongodbforms==0.3
mongoengine==0.8.7
oauthlib==0.7.2
pymongo==2.7.2
python-dateutil==2.2
python-openid==2.2.5
python-social-auth==0.2.1
pytils==0.3
pytz==2014.7
pyzmq==14.3.1
requests==2.4.1
requests-oauthlib==0.4.2
six==1.7.3
sqlparse==0.1.11
statprof==0.1.2
webassets==0.10
wsgiref==0.1.2
xmltodict==0.9.0
